requirements.txt
pybit
pandas
numpy
scipy
python-dotenv
sqlalchemy
loguru
openpyxl